requirements.txt
fastapi[standard]~=0.135.2
PyJWT[crypto]~=2.12.1
cryptography~=46.0.6
sqlmodel~=0.0.37
alembic~=1.18.4
pyyaml~=6.0.3
cosmotech-api~=5.0.1
psycopg2-binary~=2.9.11
